Removed a couple of suprious int casts in AudioFormatReaderSource

+++ modules/juce_audio_formats/format/juce_AudioFormatReaderSource.cpp
@@ -54,8 +54,8 @@ void AudioFormatReaderSource::getNextAudioBlock (const AudioSourceChannelInfo& i
 
         if (looping)
         {
-            const int newStart = (int) (start % (int) reader->lengthInSamples);
-            const int newEnd = (int) ((start + info.numSamples) % (int) reader->lengthInSamples);
+            const int64 newStart = start % reader->lengthInSamples;
+            const int64 newEnd = (start + info.numSamples) % reader->lengthInSamples;
 
             if (newEnd > newStart)
             {
@@ -64,7 +64,7 @@ void AudioFormatReaderSource::getNextAudioBlock (const AudioSourceChannelInfo& i
             }
             else
             {
-                const int endSamps = (int) reader->lengthInSamples - newStart;
+                const int64 endSamps = reader->lengthInSamples - newStart;
 
                 reader->read (info.buffer, info.startSample,
                               endSamps, newStart, true, true);
